restore create_user in `fixtures.py` + create a user with restricted access in OCCHAB
